+++ blinkreceipt-demo/BlinkReceiptCustomScan/src/main/java/com/blinkreceipt/ocr/ui/CameraActivity.java
@@ -13,7 +13,6 @@ import com.blinkreceipt.ocr.R;
 import com.microblink.BitmapResult;
 import com.microblink.CameraCaptureListener;
 import com.microblink.CameraRecognizerCallback;
-import com.microblink.CameraScanActivity;
 import com.microblink.Media;
 import com.microblink.RecognizerResult;
 import com.microblink.RecognizerView;
@@ -171,8 +170,8 @@ public class CameraActivity extends AppCompatActivity implements CameraRecognize
         finishScan.setEnabled(true);
 
         setResult(Activity.RESULT_OK, new Intent()
-                .putExtra(CameraScanActivity.DATA_EXTRA, results)
-                .putExtra(CameraScanActivity.MEDIA_EXTRA, media));
+                .putExtra(MainActivity.DATA_EXTRA, results)
+                .putExtra(MainActivity.MEDIA_EXTRA, media));
 
         finish();
     }
@@ -186,7 +185,7 @@ public class CameraActivity extends AppCompatActivity implements CameraRecognize
 
     @Override
     public void onRecognizerResultsChanged(@NonNull RecognizerResult result) {
-        Timberland.d("results: " + result.toString());
+        Timberland.d("results: " + result);
     }
 
     @Override